src/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
